refactor: remove connection & hub from Debugger
